Remove merge markers from BudgetContext
